Brings the nda-based overhaul (681 commits since the 0362e26 fork point,
Oct 2025) onto main. overhaul is the target state throughout: its layout,
its C++ rewrite, its functional tests.

overhaul had already absorbed 129 of main's 218 post-fork commits by hand,
and independently reimplemented nearly all of the rest:

  - walkerTypeIsConvertible/walkerDimsAreConvertible supersede main's
    WALKER_TYPES_name and the spin-symmetry guard (1506d26)
  - the equil_multiplier divisibility check is already in
    BackPropagatedEstimator.hpp (8c690ae, baa4f23)
  - coqui Hamiltonian reading is present across all four Hamiltonians
    (3e2cd9c, 57ab612)
  - tests/test_observables_reference.cpp covers 9 cases against main's
    1-case scaffold (be44931, f1046b1)
  - the boost eps_tolerance GPU fix (da11b12) is moot: boost is gone,
    Discrete_GeneralUJ uses in-tree math::bisect

Resolutions:
  - overhaul: all C++, docs, and afqmctools conflicts; tests/functional
    wholesale; CMakeLists.txt (the GCC>=13 CUDA guard from f5f693f is
    obsolete and deliberately dropped)
  - main: .gitmodules (https AutoHF URL, ee55051) and CONTRIBUTORS.md
  - main's only surviving feature is the momentum distribution (fbf5df5):
    rhonk.py, visualize.py, and _orbitals_to_dense_grid in qe_utils.py

FULLYPOLARIZED stays removed (overhaul PR #123); config.h emits a
migration error for old dims[3]==4 wavefunctions.

Dropped as superseded: utils/tests/functional/ (-> tests/functional/),
utils/docs/ (-> docs/afqmctools/), examples/ (-> docs/examples/),
src/AFQMC/**/tests/ (-> tests/), cmake/{GNU,Clang}Compilers.cmake,
docker/Dockerfile_naive_cuda, logo.webp, utils/tests/Readme.md.

The merged tree differs from origin/overhaul in exactly five files.

+++ utils/afqmctools/utils/io.py
@@ -234,15 +234,15 @@ def write_csr(f,csr_array:sps.csr_array,prefix):
     )
     f.create_dataset(
         name=prefix+'/jdata_',
-        data=csr_array.indices
+        data=csr_array.indices.astype(np.int32,copy=False)
     )
     f.create_dataset(
         name=prefix+'/pointers_begin_',
-        data=csr_array.indptr[:-1]
+        data=csr_array.indptr[:-1].astype(np.int32,copy=False)
     )
     f.create_dataset(
         name=prefix+'/pointers_end_',
-        data=csr_array.indptr[1:]
+        data=csr_array.indptr[1:].astype(np.int32,copy=False)
     )
 
 
@@ -324,7 +324,7 @@ def write_model_hamiltonian(
 
         f.create_dataset(
             'Hamiltonian/dims',
-            data = np.array([0, 0, 0, hamiltonian.nsites*hamiltonian.nbands , nup, ndn, 0, 0])
+            data = np.array([0, 0, 0, hamiltonian.nsites*hamiltonian.nbands , nup, ndn, 0, 0], dtype=np.int64)
         )
         # write Energies!
         f.create_dataset(
